Implementado módulo listeners
-- Implementado um módulo para conter todos os event listeners do módulo externo
-- Retirado listeners do módulo connection, ele agora retransmite todos os listeners para o event listeners interno
